perf(直播中心): 完善播放直播逻辑
1.增加获取直播间播放详情的接口
2.完善流服务器分配结构 增加播放流url字段

+++ ashe-live-business/live-center/src/main/java/cn/oneplustow/lc/service/IStreamServerService.java
@@ -43,7 +43,7 @@ public interface IStreamServerService {
      * 对指定IP的流服务器进行心跳检测
      * @param id 流服务器ID
      */
-    boolean heartBeatDetectionById(Integer id);
+    boolean heartBeatDetectionById(Long id);
 
     /**
      *
